Return from requests early
Because of the potential for logging and metrics operations to be slow or take time (many requests, for example, against different URLs will cause responses to hang momentarily waiting for a mutex lock) we want to do this stuff out of band. This means that a
client connection will close as soon as it is ready to be closed.

Currently this connection will stay open, even though all of the response data has been sent.